// packages/api-services/src/gen/_models/typeKind.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Webitel API
 * OpenAPI spec version: 24.04.0
 */

/**
 * Kind of primitive data types.
 *
 *  - none: option allow_alias = true;
 *  - list: [array]
 *  - int: int32
 *  - uint: uint32
 *  - float: float32
 *  - datetime: date &| time
 */
export type TypeKind = (typeof TypeKind)[keyof typeof TypeKind];

export const TypeKind = {
	None: 'none',
	List: 'list',
	Bool: 'bool',
	Int: 'int',
	Int32: 'int32',
	Int64: 'int64',
	Uint: 'uint',
	Uint32: 'uint32',
	Uint64: 'uint64',
	Float: 'float',
	Float32: 'float32',
	Float64: 'float64',
	Binary: 'binary',
	Lookup: 'lookup',
	String: 'string',
	Richtext: 'richtext',
	Datetime: 'datetime',
	Duration: 'duration',
} as const;